refactor: tests in internal module
