+++ .github/workflows/tests.yml
@@ -22,10 +22,7 @@ jobs:
       - name: Prepare environment for integration tests
         run: |
           rustup target add wasm32-unknown-unknown
-          cd src/integration-tests
           ./scripts/download-pocket-ic.sh
           ./scripts/build-test-canister.sh
       - name: Run integration tests
-        run: cargo test -p integration-tests -- --test-threads 1
-        env:
-          POCKET_IC_BIN: "./src/integration-tests/bin/pocket-ic"
+        run: POCKET_IC_BIN="$(pwd)/bin/pocket-ic" cargo test --package ic-websocket-cdk --lib -- tests::integration_tests --test-threads 1
