+++ bin/test_canister.wasm
@@ -1,0 +1,1 @@
+/home/eulero/Documents/omnia/ic-websocket-cdk-rs/target/wasm32-unknown-unknown/release/test_canister.wasm
