+++ src/ic-websocket-cdk/src/tests/integration_tests/a_ws_open.rs
@@ -1,12 +1,12 @@
 use std::ops::Deref;
 
-use candid::Principal;
-use ic_websocket_cdk::{
+use crate::{
     CanisterOutputMessage, CanisterWsGetMessagesArguments, CanisterWsGetMessagesResult,
     CanisterWsOpenArguments, CanisterWsOpenResult, ClientKey, WebsocketServiceMessageContent,
 };
+use candid::Principal;
 
-use crate::utils::{
+use super::utils::{
     actor::{ws_get_messages::call_ws_get_messages, ws_open::call_ws_open},
     clients::{generate_random_client_nonce, CLIENT_1, CLIENT_1_KEY, GATEWAY_1},
     messages::get_service_message_content_from_canister_message,
