+++ src/ic-websocket-cdk/src/tests/integration_tests/utils/mod.rs
@@ -9,12 +9,17 @@ pub mod test_env;
 pub mod wasm;
 
 pub fn bin_folder_path() -> PathBuf {
+    println!("manifest dir: {:?}", std::env::var("CARGO_MANIFEST_DIR"));
     let mut file_path = PathBuf::from(
         std::env::var("CARGO_MANIFEST_DIR")
             .expect("Failed to read CARGO_MANIFEST_DIR env variable"),
     );
+    file_path.pop();
+    file_path.pop();
     file_path.push("bin");
     file_path
+    // println!("{:?}", std::env::current_dir());
+    // PathBuf::from("./bin")
 }
 
 /// Returns the current timestamp in nanoseconds.
