+++ src/ic-websocket-cdk/src/tests/mod.rs
@@ -1,0 +1,4 @@
+#![cfg(test)]
+
+mod integration_tests;
+mod unit_tests;
