--- src/integration-tests/bin/test_canister.wasm
@@ -1,1 +1,0 @@
-../../../target/wasm32-unknown-unknown/release/test_canister.wasm
